lib/model: Refactor out scanning method from rwfolder.Serve loop
